+++ readme.md
@@ -9,6 +9,6 @@
 - [x] 함수(또는 메서드)의 길이가 10라인을 넘어가지 않도록 구현한다.
 - [x] 함수(또는 메서드)가 한 가지 일만 하도록 최대한 작게 만들어라.
 ### 2단계
-- [ ] 로또 당첨 번호를 받아 일치한 번호 수에 따라 당첨 결과를 보여준다.
-- [ ] 모든 원시 값과 문자열을 포장한다.
-- [ ] 일급 컬렉션을 쓴다.
+- [x] 로또 당첨 번호를 받아 일치한 번호 수에 따라 당첨 결과를 보여준다.
+- [x] 모든 원시 값과 문자열을 포장한다.
+- [x] 일급 컬렉션을 쓴다.
