+++ src/main/java/util/RandomNumberGenerator.java
@@ -3,10 +3,11 @@ package util;
 import java.util.Random;
 
 public class RandomNumberGenerator implements NumberGenerator{
-    public static final int BOUND = 45;
+    public static final int LOTTO_MAX_VALUE = 45;
+    public static final int ADJUSTED_VALUE = 1;
     Random random = new Random();
 
     public int generate(){
-        return random.nextInt(BOUND) + 1;
+        return random.nextInt(LOTTO_MAX_VALUE) + ADJUSTED_VALUE;
     }
 }
